deleted class RestAPITest and class RestAPILinkTest abd moved tests from deleted classes to FooterTest class
